Gogs: Fix data persistence
Gogs stores all git repositories in /data/git/gogs-repositories,
so they are lost after each restart. This results in error 500 when
you try to open the repository.

Also you can find generated SSH keys in /data/ssh that are being lost too.

This commit changes the gogs-data volume to point to /data to achieve
data persistence and to avoid creation of dangling volumes.